+++ src/careamics/lvae_training/dataset/utils/empty_patch_fetcher.py
