+++ src/careamics/lvae_training/dataset/utils/index_switcher.py
